+++ src/main/java/com/Medicare/security/jwt/UserDetailsImpl.java
@@ -14,7 +14,7 @@ import java.util.stream.Collectors;
 public class UserDetailsImpl implements UserDetails {
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private Integer id;
     private String username;
     private String email;
     private String fullName;
@@ -24,7 +24,7 @@ public class UserDetailsImpl implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserDetailsImpl(Long id, String username, String email, String password, String fullName
+    public UserDetailsImpl(Integer id, String username, String email, String password, String fullName
                              , Collection<? extends GrantedAuthority> authorities
                                ) {
             this.id = id;
@@ -52,7 +52,7 @@ public class UserDetailsImpl implements UserDetails {
                 );
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
